+++ CHANGELOG.md
@@ -5,11 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
 ### Added
 ### Changed
-- [issue/7] (https://github.com/podaac/virtualzarr-gen/issues/7): 
-### Deprecated
+- [issue/7] (https://github.com/podaac/virtualzarr-gen/issues/7): Update script for L2 data
+- Added aws bucket sync lambda
+### Deprecated 
 ### Removed
 ### Fixed
 ### Security